+++ src/main/java/com/depromeet/domain/auth/dto/request/MemberRegisterRequest.java
@@ -1,4 +1,4 @@
-package com.depromeet.domain.auth.dto;
+package com.depromeet.domain.auth.dto.request;
 
 public record MemberRegisterRequest(String nickname) {
     // TODO: Add validation
